+++ include/uv.h
@@ -45,18 +45,18 @@ extern "C" {
 # define UV_EXTERN /* nothing */
 #endif
 
-#include "uv-errno.h"
-#include "uv-version.h"
+#include "uv/errno.h"
+#include "uv/version.h"
 #include <stddef.h>
 #include <stdint.h>
 
 #if defined(_WIN32)
-# include "uv-win.h"
+# include "uv/win.h"
 # if !defined(BUILDING_UV_SHARED)
 #   include "io.h" /* this header is not correct in a shared library environment */
 # endif
 #else
-# include "uv-unix.h"
+# include "uv/unix.h"
 #endif
 
 /* Expand this list if necessary. */
