+++ libuv.nsi
@@ -49,11 +49,11 @@ Section "Files" SecInstall
 
 	SetOutPath "$INSTDIR\include"
 	File "include\uv.h"
-	File "include\uv-errno.h"
-	File "include\uv-threadpool.h"
-	File "include\uv-version.h"
-	File "include\uv-win.h"
-	File "include\tree.h"
+	File "include\uv\errno.h"
+	File "include\uv\threadpool.h"
+	File "include\uv\version.h"
+	File "include\uv\win.h"
+	File "include\uv\tree.h"
 
 	WriteUninstaller "$INSTDIR\Uninstall.exe"
 	WriteRegStr HKLM "Software\Microsoft\Windows\CurrentVersion\Uninstall\libuv-${ARCH}-${VERSION}" "DisplayName" "libuv-${ARCH}-${VERSION}"
@@ -73,11 +73,11 @@ Section "Uninstall"
 	Delete "$INSTDIR\README.md"
 
 	Delete "$INSTDIR\include\uv.h"
-	Delete "$INSTDIR\include\uv-errno.h"
-	Delete "$INSTDIR\include\uv-threadpool.h"
-	Delete "$INSTDIR\include\uv-version.h"
-	Delete "$INSTDIR\include\uv-win.h"
-	Delete "$INSTDIR\include\tree.h"
+	Delete "$INSTDIR\include\uv\errno.h"
+	Delete "$INSTDIR\include\uv\threadpool.h"
+	Delete "$INSTDIR\include\uv\version.h"
+	Delete "$INSTDIR\include\uv\win.h"
+	Delete "$INSTDIR\include\uv\tree.h"
 
 	Delete "$INSTDIR\Uninstall.exe"
 	RMDir "$INSTDIR"
