+++ src/unix/android-ifaddrs.c
@@ -23,7 +23,7 @@ ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include "android-ifaddrs.h"
+#include "uv/android-ifaddrs.h"
 #include "uv-common.h"
 
 #include <string.h>
