+++ src/unix/pthread-barrier.c
@@ -13,7 +13,7 @@ WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
 ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 */
-#include "pthread-barrier.h"
+#include "uv/pthread-barrier.h"
 
 #include <stdlib.h>
 #include <assert.h>
